changed to redux toolkit instead of redux core. changed to taskSlice
