+++ package.json
@@ -13,26 +13,25 @@
   },
   "dependencies": {
     "@react-navigation/bottom-tabs": "^6.5.9",
+    "@react-navigation/material-top-tabs": "^6.6.4",
     "@react-navigation/native": "^6.1.8",
     "@react-navigation/native-stack": "^6.9.14",
     "@react-navigation/stack": "^6.3.18",
+    "@reduxjs/toolkit": "^1.9.7",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "expo": "~49.0.13",
+    "expo-av": "~13.4.1",
     "expo-splash-screen": "~0.20.5",
     "expo-status-bar": "~1.6.0",
+    "lottie-react-native": "5.1.6",
     "prettier": "^3.0.3",
     "react": "^18.2.0",
     "react-native": "^0.72.5",
     "react-native-paper": "^5.10.6",
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
-    "react-redux": "^8.1.3",
-    "redux": "^4.2.1",
-    "lottie-react-native": "5.1.6",
-    "expo-av": "~13.4.1",
-    "@react-navigation/material-top-tabs": "^6.6.4",
-    "expo-av": "~13.4.1"
+    "react-redux": "^8.1.3"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
