--- store/tasks/taskActions.ts
@@ -1,8 +1,0 @@
-import { Task } from "../../types";
-
-export interface AddTaskAction {
-  type: "ADD_TASK";
-  payload: Task;
-}
-
-export type TaskAction = AddTaskAction;
